- Fixing routing based off discussion on github issue - Moving to routes without a parent by appending an "_" to the directory name and fixing the route IDs that are used
